#35 댓글 수정 삭제

+++ src/main/resources/static/js/articles/post_detail.js
@@ -1,5 +1,5 @@
 let url_list = window.location.href.split('/')
-const id = url_list[url_list.length - 1]
+const id = url_list[url_list.length - 1].replace(/[^0-9.]/g, '')
 
 $(document).ready(function () {
     if (localStorage.getItem('token')) {
@@ -51,6 +51,52 @@ function show_post(id) {
     })
 }
 
+function update_img() {
+    var reader = new FileReader();
+    reader.onload = function (event) {
+        var img = document.createElement("upfile");
+        img.setAttribute("src", event.target.result);
+        document.querySelector("img#ucontent-img").appendChild(img);
+    };
+    reader.readAsDataURL(event.target.files[0]);
+}
+
+function update_post() {
+    if (confirm("수정 하시겠습니까?") == false) {
+        $('#modalClose').click();
+    } else {
+        let comment = $('#update-content').val();
+        let title = $('#update-title').val();
+        let file = null;
+
+        let fileInput = document.getElementsByClassName("up");
+
+        if (fileInput.length > 0) {
+            file = $('#upfile')[0].files[0]
+        }
+
+        const formData = new FormData();
+
+        formData.append("title", title);
+        formData.append("content", content);
+        formData.append("img", file);
+        formData.append("idx", id);
+        $.ajax({
+            type: "PUT",
+            url: `/posts/detail`,
+            processData: false,
+            contentType: false,
+            data: formData,
+            success: function (response) {
+                window.location.reload();
+            },
+            error: function (request, status, error) {
+                alert(error);
+            }
+        });
+    }
+
+}
 
 function delete_post() {
     const idx = $("#idx").val();
@@ -82,13 +128,28 @@ function show_comment(comments) {
                                 <div class="card-header bg-light">
                                     <i class="fa fa-comment fa"></i> 작성자: ${e.user['username']}
                                     <input hidden value="${e['idx']}">
+                                        <div class="h-auto dropdown">
+                                          <a class="btn btn-secondary dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-bs-toggle="dropdown" aria-expanded="false">
+                                            Edit
+                                          </a>
+                                           
+                                          <ul class="dropdown-menu" aria-labelledby="dropdownMenuLink">
+                                            <li><a class="dropdown-item" href="#" onclick="comment_update_input(${e['idx']})">수정</a></li>
+                                            <li><a class="dropdown-item" href="#" onclick="comment_delete(${e['idx']})">삭제</a></li>
+                                          </ul>
+                                        </div>
                                 </div>
                                 <div class="card-body">
                                     <ul class="list-group list-group-flush">
                                         <li class="list-group-item">
-                                            <div class="comment_wrote">내용: ${e.comment}</div>
-                                            <button onclick="comment_update()" type="button" class="btn btn-dark mt-3">수정</button>
-                                            <button onclick="comment_delete()" type="button" class="btn btn-dark mt-3">삭제</button>
+                                            <div class="comment_wrote">${e.comment}</div>
+                                               <!--수정 댓글-->
+                                               
+                                             <div id="comment_input_${e['idx']}" class="d-none">
+                                                <input id="comment_upvalue_${e['idx']}" type="text"  class="form-control">
+                                                <button onclick="comment_update(${e['idx']})" type="button" class="h-auto btn btn-dark mt-3">수정</button>
+                                             </div>
+                                             
                                         </li>
                                     </ul>
                                 </div>
@@ -99,6 +160,7 @@ function show_comment(comments) {
     $(`#comment_list`).html(comment_text)
 }
 
+//댓글 업로드
 function comment_upload() {
     const comment_input = $("#comment_content").val();
     if (comment_input.length == 0) {
@@ -120,18 +182,32 @@ function comment_upload() {
             const arr_comment = response.reverse();
             arr_comment.forEach((e) => {
                 comment_text += `
-                            &nbsp;
                             <div class="card mb-2">
                                 <div class="card-header bg-light">
                                     <i class="fa fa-comment fa"></i> 작성자: ${e.user['username']}
                                     <input hidden value="${e['idx']}">
-                               </div>
+                                        <div class="h-auto dropdown">
+                                          <a class="btn btn-secondary dropdown-toggle" href="#" role="button" id="dropdownMenuLink" data-bs-toggle="dropdown" aria-expanded="false">
+                                            Edit
+                                          </a>
+                                           
+                                          <ul class="dropdown-menu" aria-labelledby="dropdownMenuLink">
+                                            <li><a class="dropdown-item" href="#" onclick="comment_update_input(${e['idx']})">수정</a></li>
+                                            <li><a class="dropdown-item" href="#" onclick="comment_delete(${e['idx']})">삭제</a></li>
+                                          </ul>
+                                        </div>
+                                </div>
                                 <div class="card-body">
                                     <ul class="list-group list-group-flush">
                                         <li class="list-group-item">
-                                            <div class="comment_wrote">내용: ${e.comment}</div>
-                                            <button onclick="comment_update()" type="button" class="btn btn-dark mt-3">수정</button>
-                                            <button onclick="comment_delete()" type="button" class="btn btn-dark mt-3">삭제</button>
+                                            <div class="comment_wrote">${e.comment}</div>
+                                               <!--수정 댓글-->
+                                               
+                                             <div id="comment_input_${e['idx']}" class="d-none">
+                                                <input id="comment_upvalue_${e['idx']}" type="text"  class="form-control">
+                                                <button onclick="comment_update(${e['idx']})" type="button" class="h-auto btn btn-dark mt-3">수정</button>
+                                             </div>
+                                             
                                         </li>
                                     </ul>
                                 </div>
@@ -148,41 +224,53 @@ function comment_upload() {
     });
 }
 
+function comment_update_input(id){
+    $("#comment_input_"+id).toggleClass('d-none')
+}
+//댓글 수정
+function comment_update(id) {
+    const result = confirm("수정 하시겠습니까?");
 
-function comment_delete() {
-    alert("삭제 기능 입니다!")
-    const idx = $("#idx").val();
-    // comment idx 찾아줘야함
-    $.ajax({
-        type: "DELETE",
-        url: `/comment/${idx}`,
-        data: {
-            id_give: idx,
-        },
-        success: function (response) {
-            console.log(response['result'])
-        },
-        error: function (request, status, error) {
-            console.log(error);
+    if(result){
+        data ={
+            commentid:id,
+            comment:$("#comment_upvalue_"+id).val()
         }
-    })
+        $.ajax({
+            type: "PUT",
+            url: `/posts/comment`,
+            data: JSON.stringify(data),
+            contentType: 'application/json; charset=utf-8',
+            success: function (response) {
+                window.location.reload();
+            },
+            error: function (request, status, error) {
+                console.log(error);
+            }
+        })
+    }
+    else{
+        comment_update_input(id)
+    }
+
 }
 
-function comment_update() {
-    alert("수정 버튼 입니다!")
-    const idx = $("#idx").val();
-    // comment idx 찾아줘야함
-    $.ajax({
-        type: "PUT",
-        url: `/comment/${idx}`,
-        data: {
-            id_give: idx,
-        },
-        success: function (response) {
-            console.log(response['result'])
-        },
-        error: function (request, status, error) {
-            console.log(error);
-        }
-    })
+//댓글 삭제
+function comment_delete(id) {
+    const result = confirm("댓글을 삭제 하시겠습니까?");
+
+    if (result) {
+        $.ajax({
+            type: "DELETE",
+            url: `/posts/comment`,
+            contentType: 'application/json; charset=utf-8',
+            data: JSON.stringify({commentid:id}),
+            success: function (response) {
+                window.location.reload();
+            },
+            error: function (request, status, error) {
+                console.log(error);
+            }
+        })
+    }
 }
